requirements.txt
absl-py==0.8.1
astor==0.8.0
certifi==2019.11.28
cloudpickle==1.2.2
future==0.18.2
gast==0.2.2
google-pasta==0.1.8
grpcio==1.16.1
gym==0.15.6
h5py==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
numpy==1.18.1
opt-einsum==3.1.0
protobuf==3.11.2
pyglet==1.4.10
PyYAML==5.3
scipy==1.4.1
six==1.14.0
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.0
termcolor==1.1.0
Werkzeug==0.16.1
wrapt==1.11.2
